requirements.txt
pathlib
fastapi
uvicorn[standard]
httpx
python-dotenv
